dropped the Updater code entirely for the 1.0 release as we don't need it anymore moved the cz.posvic activity test into org.mozilla.mozstumbler
