+++ android/src/test/java/org/mozilla/mozstumbler/client/MainActivityTest.java
@@ -1,0 +1,29 @@
+package org.mozilla.mozstumbler.client;
+
+import org.junit.Before;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.robolectric.Robolectric;
+import org.robolectric.RobolectricTestRunner;
+import org.robolectric.annotation.Config;
+
+import static org.junit.Assert.assertNotNull;
+
+@Config(emulateSdk = 18)
+@RunWith(RobolectricTestRunner.class)
+public class MainActivityTest {
+
+    private MainActivity activity;
+
+    @Before
+    public void setUp() throws Exception {
+        activity = Robolectric.buildActivity(MainActivity.class).create().start().resume().get();
+    }
+
+    @Test
+    public void activityShouldNotBeNull() {
+        assertNotNull(activity);
+    }
+
+
+}
